requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
sqlparse>=0.5.0
rich>=13.0.0
python-dotenv>=1.0.0
